Request for comments
1. Created tests for pkg/router. Problems exists and

- some of them are problem of test environment
- and some are problem discovered by tests

Discovered problems:

- `Example_setupHandlers_confirmLoop` in **most** (but not all) cases
finishes with "chacha20poly1305: message authentication failed"
- `logging.LevelFromString` + `logging.SetLevel`  causes data races
- ProcManager.RunProc - sometimes panics
- periodical problems arises **after** the test - Close/Stop/TearDown
logic is not very clear

2. Created mini-framework for creating environment, running tests in
them, stopping/tearing down them

Typical test looks like:

```golang
	env := &TEnv{}
	_, err := env.RunAsExample(true,
		GenerateKeys(),
		AddTransportManagers(),
		AddProcManagerAndRouter(),
		AddSetupHandlersEnv(),
		// Preparation for confirmLoop
		AddRules(),
		AddLoopData(),
		AddAppAndRunProc("/bin", "sh"), // This step is unstable
		//confirmLoop entrails
		CheckRulesAndPorts(),
		ConfirmLoopAndFinish(),
	)
```

Idea for such test composition came from Rob Pike: ["Self-referential functions and the design of options"]
(https://commandcenter.blogspot.com/2014/01/self-referential-functions-and-design.html)

Look at `pkg/router/tenv_test.go` and `pkg/router/cfgsteps_test.go`
